+++ io/src/main/java/org/n52/io/quantity/QuantityIoFactory.java
@@ -32,17 +32,20 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+
+import org.n52.io.Constants;
 import org.n52.io.IoFactory;
 import org.n52.io.IoHandler;
 import org.n52.io.IoProcessChain;
-import org.n52.io.MimeType;
-import org.n52.io.quantity.csv.QuantityCsvIoHandler;
-import org.n52.io.quantity.format.FormatterFactory;
-import org.n52.io.quantity.generalize.GeneralizingQuantityService;
-import org.n52.io.quantity.img.ChartIoHandler;
-import org.n52.io.quantity.img.MultipleChartsRenderer;
-import org.n52.io.quantity.report.PDFReportGenerator;
+import org.n52.io.csv.quantity.QuantityCsvIoHandler;
+import org.n52.io.format.ResultTimeFormatter;
+import org.n52.io.format.quantity.FormatterFactory;
+import org.n52.io.generalize.quantity.GeneralizingQuantityService;
+import org.n52.io.img.quantity.ChartIoHandler;
+import org.n52.io.img.quantity.MultipleChartsRenderer;
+import org.n52.io.report.quantity.PDFReportGenerator;
 import org.n52.io.request.IoParameters;
+import org.n52.io.request.Parameters;
 import org.n52.io.response.dataset.DataCollection;
 import org.n52.io.response.dataset.quantity.QuantityData;
 import org.n52.io.response.dataset.quantity.QuantityDatasetOutput;
@@ -51,12 +54,12 @@ import org.n52.series.spi.srv.DataService;
 
 public final class QuantityIoFactory extends IoFactory<QuantityData, QuantityDatasetOutput, QuantityValue> {
 
-    private static final List<MimeType> SUPPORTED_MIMETYPES = Arrays.asList(
-            new MimeType[] {
-                MimeType.TEXT_CSV,
-                MimeType.IMAGE_PNG,
-                MimeType.APPLICATION_ZIP,
-                MimeType.APPLICATION_PDF,
+    private static final List<Constants.MimeType> SUPPORTED_MIMETYPES = Arrays.asList(
+            new Constants.MimeType[] {
+                Constants.MimeType.TEXT_CSV,
+                Constants.MimeType.IMAGE_PNG,
+                Constants.MimeType.APPLICATION_ZIP,
+                Constants.MimeType.APPLICATION_PDF,
             });
 
     @Override
@@ -64,64 +67,69 @@ public final class QuantityIoFactory extends IoFactory<QuantityData, QuantityDat
         return new IoProcessChain<QuantityData>() {
             @Override
             public DataCollection<QuantityData> getData() {
-                final boolean generalize = getParameters().isGeneralize();
+                boolean generalize = getParameters().isGeneralize();
                 DataService<QuantityData> dataService = generalize
                         ? new GeneralizingQuantityService(getDataService())
                         : getDataService();
-                return dataService.getData(getRequestParameters());
+                return dataService.getData(getParameters());
             }
 
             @Override
             public DataCollection<?> getProcessedData() {
-                String format = getParameters().getFormat();
-                return FormatterFactory.createFormatterFactory(format).create().format(getData());
+                return getParameters().shallClassifyByResultTimes()
+                        ? new ResultTimeFormatter<QuantityData>().format(getData())
+                        : createFormatter().create().format(getData());
+            }
+
+            private FormatterFactory createFormatter() {
+                return FormatterFactory.createFormatterFactory(getParameters());
             }
         };
     }
 
     @Override
     public boolean isAbleToCreateHandlerFor(String outputMimeType) {
-        return MimeType.isKnownMimeType(outputMimeType)
-                && supportsMimeType(MimeType.toInstance(outputMimeType));
+        return Constants.MimeType.isKnownMimeType(outputMimeType)
+                && supportsMimeType(Constants.MimeType.toInstance(outputMimeType));
     }
 
     @Override
     public Set<String> getSupportedMimeTypes() {
         HashSet<String> mimeTypes = new HashSet<>();
-        for (MimeType supportedMimeType : SUPPORTED_MIMETYPES) {
+        for (Constants.MimeType supportedMimeType : SUPPORTED_MIMETYPES) {
             mimeTypes.add(supportedMimeType.getMimeType());
         }
         return mimeTypes;
     }
 
-    private static boolean supportsMimeType(MimeType mimeType) {
+    private static boolean supportsMimeType(Constants.MimeType mimeType) {
         return SUPPORTED_MIMETYPES.contains(mimeType);
     }
 
     @Override
     public IoHandler<QuantityData> createHandler(String outputMimeType) {
         IoParameters parameters = getParameters();
-        MimeType mimeType = MimeType.toInstance(outputMimeType);
-        if (mimeType == MimeType.IMAGE_PNG) {
+        Constants.MimeType mimeType = Constants.MimeType.toInstance(outputMimeType);
+        if (mimeType == Constants.MimeType.IMAGE_PNG) {
             return createMultiChartRenderer(mimeType);
-        } else if (mimeType == MimeType.APPLICATION_PDF) {
+        } else if (mimeType == Constants.MimeType.APPLICATION_PDF) {
             ChartIoHandler imgRenderer = createMultiChartRenderer(mimeType);
             PDFReportGenerator reportGenerator = new PDFReportGenerator(
-                    getRequestParameters(),
+                    parameters,
                     createProcessChain(),
                     imgRenderer);
             reportGenerator.setBaseURI(getBasePath());
             return reportGenerator;
-        } else if (mimeType == MimeType.TEXT_CSV || mimeType == MimeType.APPLICATION_ZIP) {
+        } else if (mimeType == Constants.MimeType.TEXT_CSV || mimeType == Constants.MimeType.APPLICATION_ZIP) {
             QuantityCsvIoHandler handler = new QuantityCsvIoHandler(
-                    getRequestParameters(),
+                    parameters,
                     createProcessChain(),
                     getMetadatas());
-            handler.setTokenSeparator(parameters.getOther("tokenSeparator"));
+            handler.setTokenSeparator(parameters.getOther(Parameters.TOKEN_SEPARATOR));
 
-            boolean zipOutput = parameters.getAsBoolean(MimeType.APPLICATION_ZIP.name());
-            handler.setZipOutput(zipOutput || mimeType == MimeType.APPLICATION_ZIP);
-            boolean byteOderMark = Boolean.parseBoolean(parameters.getOther("bom"));
+            boolean zipOutput = parameters.getAsBoolean(Parameters.ZIP, false);
+            handler.setZipOutput(zipOutput || mimeType == Constants.MimeType.APPLICATION_ZIP);
+            boolean byteOderMark = Boolean.parseBoolean(parameters.getOther(Parameters.BOM));
             handler.setIncludeByteOrderMark(byteOderMark);
             return handler;
         }
@@ -131,9 +139,9 @@ public final class QuantityIoFactory extends IoFactory<QuantityData, QuantityDat
         throw exception;
     }
 
-    private MultipleChartsRenderer createMultiChartRenderer(MimeType mimeType) {
+    private MultipleChartsRenderer createMultiChartRenderer(Constants.MimeType mimeType) {
         MultipleChartsRenderer chartRenderer = new MultipleChartsRenderer(
-                getRequestParameters(),
+                getParameters(),
                 createProcessChain(),
                 createContext());